Rename and add test logger
