--- app/utils/logger/bunyan-logger.ts
@@ -1,12 +1,0 @@
-import { remote } from "electron";
-let logger = (remote.getCurrentWindow() as any).logger;
-if (!logger) {
-    logger = {
-        trace: () => null,
-        error: () => null,
-        log: () => null,
-        warn: () => null,
-        info: () => null,
-    };
-}
-export default logger;
