+++ src/@batch-flask/utils/index.ts
@@ -1,4 +1,4 @@
-export * from "./logger";
+export * from "./logging";
 export * from "./date-utils";
 export * from "./form-utils";
 export * from "./secure-utils";
