+++ src/@batch-flask/utils/logging/index.ts
@@ -2,7 +2,11 @@ import { Logger } from "./base-logger";
 
 // tslint:disable:no-var-requires
 let _logger: any;
-if (process.env.RENDERER) {
+
+if (process.env.NODE_ENV === "test") {
+    const { TestLogger } = require("./test-logger");
+    _logger = new TestLogger();
+} else if (process.env.RENDERER) {
     const { RendererLogger } = require("./renderer-logger");
     _logger = new RendererLogger();
 } else {
