+++ src/@batch-flask/utils/logging/test-logger.ts
@@ -1,0 +1,19 @@
+import { Logger } from "./base-logger";
+
+export class TestLogger implements Logger {
+    public trace(message: string) {
+        // Nothing
+    }
+    public debug(message: string, ...params: any[]) {
+        // Nothing
+    }
+    public info(message: string, ...params: any[]) {
+        // Nothing
+    }
+    public warn(message: string, params?: any) {
+        // Nothing
+    }
+    public error(message: string, error?: any) {
+        // Nothing
+    }
+}
