+++ src/client/logger/index.ts
@@ -1,1 +1,1 @@
-export * from "./logger-proxy";
+export * from "./loggers";
